[MOVE] : article-qa-4 / #100
파일 위치 변경
